+++ backend/profiles/views/__init__.py
@@ -1,3 +1,4 @@
+from .current_user_view import CurrentUserView
 from .friend_request_view import FriendRequestAPIView
 from .friendship_view import FriendshipAPIView
 from .joint_watchlist_view import JointWatchlistAPIView
@@ -5,9 +6,7 @@ from .matchlist_dislike_view import MatchlistDislikeAPIView
 from .matchlist_like_view import MatchlistLikeAPIView
 from .matchlist_match_view import MatchlistMatchAPIView
 from .movie_details_for_user_view import MovieDetailsForUserView
-from .user_detail_view import UserDetailView
-from .user_search_view import UserSearchView
-from .user_view import UserAPIView
+from .user_view import UserView
 from .watchlist_view import WatchlistAPIView
 
 __all__ = [
@@ -18,8 +17,7 @@ __all__ = [
     'MatchlistLikeAPIView',
     'MatchlistMatchAPIView',
     'MovieDetailsForUserView',
-    'UserDetailView',
-    'UserSearchView',
-    'UserAPIView',
+    'UserView',
+    'CurrentUserView',
     'WatchlistAPIView',
 ]
