--- backend/profiles/views/user_detail_view.py
@@ -1,18 +1,0 @@
-from django.shortcuts import get_object_or_404
-from rest_framework import generics
-from rest_framework.permissions import IsAuthenticated
-from rest_framework.response import Response
-from rest_framework.views import APIView
-
-from ..models import User
-from ..serializers import UserSerializer
-
-
-class UserDetailView(generics.RetrieveAPIView):
-    serializer_class = UserSerializer
-    permission_classes = [IsAuthenticated]
-    lookup_url_kwarg = 'user_id'
-
-    def get_object(self):
-        user = get_object_or_404(User, id=self.kwargs.get('user_id'))
-        return user
