+++ backend/profiles/views/user_view.py
@@ -1,17 +1,21 @@
+from django.shortcuts import get_object_or_404
+from rest_framework import viewsets
 from rest_framework.permissions import IsAuthenticated
-from rest_framework.response import Response
-from rest_framework.views import APIView
 
-from ..serializers import UserSerializer
+from ..models import User
+from ..serializers import SimpleUserSerializer, UserSerializer
 
 
-class UserAPIView(APIView):
+class UserView(viewsets.ModelViewSet):
+    serializer_class = UserSerializer
+    permission_classes = [IsAuthenticated]
 
-    permission_classes = (IsAuthenticated,)
+    def get_queryset(self):
+        search = self.request.GET.get('search', '')
+        return User.objects.search(search)
 
-    def get(self, request):
-        user = request.user
+    def get_serializer_class(self):
+        if self.action == 'list':
+            return SimpleUserSerializer
 
-        serializer = UserSerializer(user)
-
-        return Response(serializer.data)
+        return self.serializer_class
